+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added Proxy tile @pnicolli
 - Add toast component @sneridagh
+- Add sidebar support for components @sneridagh
 
 ### Changes